requirements.txt
discord.py
requests
asyncio
pysqlite3-binary
python-dotenv